Use Object#singleton_methods instead of Object#methods(false)
In 1.9 multiple versions of Delegator do not accept the extra argument for Object#methods, and Object#methods is not listed in the documentation for Ruby 1.9. Even though it does work we cannot hope it will be available forever. Since there is a published way to get singleton methods and it's backwards compatible it might be a good idea to just use that.